backend/apps/game/views.py
# views.py
from rest_framework import viewsets
from .models import Category, Game
from .serializers import CategorySerializer, GameSerializer

class CategoryViewSet(viewsets.ModelViewSet):
    queryset = Category.objects.all()
    serializer_class = CategorySerializer

class GameViewSet(viewsets.ModelViewSet):
    queryset = Game.objects.all()
    serializer_class = GameSerializer